requirements.txt
PyAutoGUI
opencv-python
python-dotenv
pillow
tkinter

#pip install opencv-python
#pip install -r requirements.txt